API implemented according to our DB scheme

+++ UpicResortsServer/src/main/java/com/cs6550/upicresortsserver/repositories/SkierRepository.java
@@ -3,6 +3,6 @@ package com.cs6550.upicresortsserver.repositories;
 import com.cs6550.upicresortsserver.models.Skier;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SkierRepository extends CrudRepository<Skier, Long> {
+public interface SkierRepository extends CrudRepository<Skier, Integer> {
 
 }
